chore: Split subscription logic from component, export helper fn
