+++ src/lib/index.ts
@@ -1,10 +1,11 @@
 import Breakpoints from '$lib/Breakpoints.svelte';
-import { useMediaQuery } from '$lib/internal';
+import { subscribeToQueries, useMediaQuery } from '$lib/internal.svelte';
 
 import type { BreakpointQueries, BreakpointMatch } from '$lib/types';
 
 export {
   Breakpoints as default,
+  subscribeToQueries,
   useMediaQuery,
   type BreakpointQueries,
   type BreakpointMatch
